feat: Implement ExerciseImage and VideoPlayerWidget for enhanced exercise display
