+++ lib/features/exercise/presentation/views/widgets/exercise_widget_item.dart
@@ -1,126 +1,25 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:fitness_app/core/utils/helper/extention.dart';
 import 'package:fitness_app/core/utils/helper_func/youtube_thumbnail.dart';
-import 'package:fitness_app/core/utils/theme/app_colors.dart';
 import 'package:fitness_app/features/exercise/data/models/exercise_by_prime_mover_and_difficulty_model/exercise.dart';
+import 'package:fitness_app/features/exercise/presentation/views/widgets/exercise_image.dart';
+import 'package:fitness_app/features/exercise/presentation/views/widgets/video_player_widget.dart';
 import 'package:flutter/material.dart';
-import 'package:youtube_player_flutter/youtube_player_flutter.dart';
 
-class ExerciseWidgetItem extends StatefulWidget {
+class ExerciseWidgetItem extends StatelessWidget {
   const ExerciseWidgetItem({super.key, required this.exercise});
   final Exercise exercise;
 
   @override
-  State<ExerciseWidgetItem> createState() => _ExerciseWidgetItemState();
-}
-
-class _ExerciseWidgetItemState extends State<ExerciseWidgetItem> {
-  YoutubePlayerController? _youtubeController;
-
-  void showVideoDialog() {
-    final videoId = YoutubePlayer.convertUrlToId(
-      widget.exercise.shortYoutubeDemonstrationLink ?? '',
-    );
-
-    if (videoId == null) return;
-    _youtubeController = YoutubePlayerController(
-      initialVideoId: videoId,
-      flags: const YoutubePlayerFlags(
-        // autoPlay: true,
-        mute: false,
-        // controlsVisibleAtStart: true,
-      ),
-    );
-    showDialog(
-      context: context,
-      builder: (context) {
-        return Dialog(
-          backgroundColor: Colors.transparent,
-          insetPadding: EdgeInsets.zero,
-          child: Container(
-            decoration: BoxDecoration(
-              gradient: const LinearGradient(
-                begin: Alignment.topCenter,
-                end: Alignment.bottomCenter,
-                colors: [AppColors.baseBlack, AppColors.neutral90],
-              ),
-
-              boxShadow: [
-                BoxShadow(
-                  color: const Color(0xFF000000).withOpacity(0.2),
-                  spreadRadius: 5,
-                  blurRadius: 7,
-                  offset: const Offset(0, 3),
-                ),
-              ],
-            ),
-            height: context.height * 0.40,
-            width: context.width,
-            child: YoutubePlayer(
-              controller: _youtubeController!,
-              showVideoProgressIndicator: true,
-              progressIndicatorColor: AppColors.baseWhite,
-            ),
-          ),
-        );
-      },
-    ).then((_) {
-      _youtubeController?.pause();
-      _youtubeController?.dispose();
-      _youtubeController = null;
-    });
-  }
-
-  @override
-  void dispose() {
-    _youtubeController?.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
     final thumbnailUrl = getYouTubeThumbnail(
-     widget.exercise.shortYoutubeDemonstrationLink ?? "",
+      exercise.shortYoutubeDemonstrationLink ?? "",
     );
     return Padding(
       padding: const EdgeInsets.all(8.0),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.center,
         children: [
-          Container(
-            height: 88,
-            width: 81,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(20),
-              color: Colors.grey[300],
-            ),
-            clipBehavior: Clip.antiAlias,
-            child:
-                thumbnailUrl.isNotEmpty
-                    ? CachedNetworkImage(
-                      imageUrl: thumbnailUrl,
-                      fit: BoxFit.fill,
-                      placeholder:
-                          (context, url) => const Center(
-                            child: CircularProgressIndicator.adaptive(
-                              strokeWidth: 2,
-                              constraints: BoxConstraints.tightFor(
-                                width: 20,
-                                height: 20,
-                              ),
-                              valueColor: AlwaysStoppedAnimation<Color>(
-                                AppColors.mainRed,
-                              ),
-                              backgroundColor: AppColors.mainRed,
-                            ),
-                          ),
-                      errorWidget:
-                          (context, url, error) => const Center(
-                            child: Icon(Icons.error, color: Colors.red),
-                          ),
-                    )
-                    : const Center(child: Icon(Icons.error, color: Colors.red)),
-          ),
+          ExerciseImage(thumbnailUrl: thumbnailUrl),
 
           const SizedBox(width: 16),
 
@@ -130,7 +29,7 @@ class _ExerciseWidgetItemState extends State<ExerciseWidgetItem> {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 Text(
-                  'Bench Press',
+                  exercise.exercise ?? "",
                   style: context.textTheme.titleSmall?.copyWith(
                     fontWeight: FontWeight.w500,
                   ),
@@ -144,21 +43,12 @@ class _ExerciseWidgetItemState extends State<ExerciseWidgetItem> {
               ],
             ),
           ),
-
-          GestureDetector(
-            onTap: showVideoDialog,
-            child: const CircleAvatar(
-              backgroundColor: AppColors.mainRed,
-              radius: 15,
-              child: Icon(
-                Icons.play_arrow,
-                color: AppColors.neutral90,
-                size: 30,
-              ),
-            ),
-          ),
+          VideoPlayerWidget(exercise: exercise),
         ],
       ),
     );
   }
 }
+
+
+
